Restrict installation of RAP e4 platform bundles to RAP environment
Use Require-Capability: org.eclipse.rap to make sure that the modified
RAP incarnations of these copies of the Platform bundles are only used
together with the correct RAP bundles.

During p2 installations require the existence of RWT, and prevent the
installations as soon as the SWT installable unit had been installed.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=491177
Change-Id: I1accaab38c5d0e67860c6b2267b8aa17db10ee44
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>