Ensure workbench cleanup on UIThread termination
RAP UI session is terminated by terminating the UIThread. In this case
ThreadDeath error is thrown. In order to avoid a memory leak we should
close the workbench and dispose the root context in this case too.

Wrap the call to E4Workbench#createAndRunUI in E4EntryPointFactory in
try/finally block to ensure workbench/context cleanup.

Change-Id: Icb377977036b17e7b5704098afb3726e990fd12f